refactor: Refactor AboutPage architecture
