+++ src/app/components/Layout/AboutPage/index.tsx
@@ -1,0 +1,18 @@
+import React from 'react';
+import { Section } from './Section';
+import SectionsData from './data/SectionsData';
+
+export const AboutPage: React.FunctionComponent = () => (
+	<React.Fragment>
+		{SectionsData.map((sectionData) => (
+			<Section
+				title={sectionData.title}
+				text={sectionData.text}
+				background={sectionData.background}
+				backgroundAside={sectionData.backgroundAside}
+				linkText={sectionData.linkText}
+				linkUrl={sectionData.linkUrl}
+			/>
+		))}
+	</React.Fragment>
+);
